Changes SecurityConfiguration -> SecurityConfigurationBase so that implementing applications have less of a chance of a naming conflict.
